requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
geographiclib==2.0
geopy==2.4.1
kombu==5.3.6
numpy==1.26.4
pandas==2.2.1
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
python-dotenv
psycopg2-binary